+++ src/cpskin/citizen/tests/test_vocabularies.py
@@ -31,9 +31,9 @@ class TestVocabularies(testing.BaseTestCase):
     def test_citizen_portal_types_vocabulary(self):
         """Test 'CitizensPortalTypesVocabulary'"""
         voc = self._get_vocabulary_values('cpskin.citizen.portal_types')
-        self.assertListEqual(['Document'], voc)
+        self.assertListEqual(sorted(['Document', 'News Item', 'Event']), sorted(voc))
 
     def test_citizen_creation_types_vocabulary(self):
         """Test 'CitizensAllowedCreationTypesVocabulary'"""
         voc = self._get_vocabulary_values('cpskin.citizen.allowed_creation_portal_types')
-        self.assertListEqual(['Document'], voc)
+        self.assertListEqual(sorted(['Document']), sorted(voc))